% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutterplot.R
\name{shutterplotsummary}
\alias{shutterplotsummary}
\title{Numerical Summaries of a Shutter Plot}
\usage{
shutterplotsummary(x, y, getValue = FALSE)
}
\arguments{
\item{x}{data for the explanatory/independent variable.}

\item{y}{data for the response/dependent variable.}

\item{getValue}{logical:FALSE (DEFAULT); to access the summary statistics of the shutter plot.}
}
\value{
Prints the numerical summaries in the console.
}
\description{
displays numerical summaries of a shutter plot.
}
\examples{
  data1 <- rnorm(90,10,10)
  data2 <- data1 + rnorm(90,20,10)
  shutterplotsummary(data1,data2)
}
