% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutterplot.R
\name{summary7}
\alias{summary7}
\title{Numerical values of seven-number-summary.}
\usage{
summary7(x)
}
\arguments{
\item{x}{value(s) of a variable.}
}
\value{
prints the seven-number-summary in the console.
}
\description{
prints the numerical summaries in the console.
}
\examples{
  data <- rnorm(90,90,10)
  summary7(data)
}
