% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutterplot.R
\name{summary7plot}
\alias{summary7plot}
\title{seven-number-summary}
\usage{
summary7plot(x)
}
\arguments{
\item{x}{value(s) of a variable.}
}
\value{
depicts the seven-number-summary.
}
\description{
displays the seven-number-summary for a variable.
}
\examples{
  data<- rnorm(90,90,10)
  summary7plot(data)
}
